Operator program

+++ Basic/Q34.java
@@ -2,12 +2,11 @@ import java.util.Scanner;
 class Q34{
    public static void main(String[]args){
 	   Scanner sc=new Scanner(System.in);
-	  double a=sc.nextInt();
-	  
-	  double b=sc.nextInt();
-	 
+	  double a=sc.nextDouble();
+	  double b=sc.nextDouble();
 	  double c=a+b;
+	  
+	  
 	  System.out.println(c);
-	   
    }
 }
